+++ CHANGELOG.md
@@ -4,6 +4,14 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased] Unreleased
 
+## Fixed
+
+- Check the PID file for the PHP built-in server, MySQL and Chromedriver controllers to make sure the PID maps to an actually running process.
+
+## Changed
+
+- Inherit environment variables while parsing configuration files in the `WPConfigFile` class.
+
 ## [3.7.9] 2024-11-27;
 
 ## Added